+++ workflow.md
@@ -24,7 +24,7 @@ Users can deploy, manage, and track their ERC20 token deployments.
 14. [ ] Modularize App.css
 15. [ ] Add contract methods to Contract Actions tab.
 16. [ ] Add a success\fail message on contract deployment(start with alert and continue with a component). 
-17. [ ] Add a div to nest a curved asset component and wrap the logo and text in it.
+17. [x] Add a div to nest a curved asset component and wrap the logo and text in it.
 18. [ ] Change all px props to % or similar for margins and paddings.
 19. [ ] Change Wallet Actions text to wallet hash and network and if not connect then it will have a "No Connected Wallet". 
 20. [ ] Fix Deploy route to app.jsx.
@@ -32,4 +32,4 @@ Users can deploy, manage, and track their ERC20 token deployments.
 22. [ ] Investigate walletConnect deprecation warning.
 23. [ ] Add created by catscan in the footer.
 24. [ ]
-25. [ ]
+25. [ ]
